refactor(#138) : 파일 구조 가독성 증가
