+++ src/main/java/odyssey/backend/domain/node/exception/NodeNotFoundException.java
@@ -1,5 +1,6 @@
 package odyssey.backend.domain.node.exception;
 
+import odyssey.backend.domain.node.exception.error.NodeExceptionProperty;
 import odyssey.backend.shared.exception.GlobalException;
 
 public class NodeNotFoundException extends GlobalException {
